% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{amphibians}
\alias{amphibians}
\title{Latin names of Amphibian Species}
\format{
A \code{data.frame} with 6618 observations and 10 variables.
}
\description{
Latin names of Amphibian Species
}
\details{
This dataset contains all Latin Names of Amphibians 
that were derived from the IUCN database, 2017 version.
}
