% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{amphibians_dist}
\alias{amphibians_dist}
\title{Distribution of amphibians at 0.5 degree spatial resolution}
\format{A \code{data.frame} with 792695 observations and 3 variables.}
\description{
Data with x,y coordinates for each amphibian species (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all amphibian species (on a 0.5 degree grid)
according to the IUCN range maps.
}
