% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{amphibians_dist_smallrange}
\alias{amphibians_dist_smallrange}
\title{Distribution of small ranging amphibians}
\format{
A \code{data.frame} with 10463 observations and 5 variables.
}
\description{
Data with x,y coordinates for each amphibian species that has less than 10 occurrences (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all amphibian species that have a small range (less than 10 occurrences on a 0.5 degree grid)
according to the IUCN range maps, 2017 version.
}
