% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{amphibians_endemic}
\alias{amphibians_endemic}
\title{Names of endemic amphibian Species}
\format{
A \code{data.frame} with 2700 observations and 2 variables.
}
\description{
Data with endemic amphibian species names and their range area
}
\details{
This dataset contains the name and range area of all endemic amphibian species. 
Species are classified as endemic if their range area is smaller or equal to 
the range area of the smallest ranging 15% of all species. 
Distribution data was derived from IUCN range maps, 2017 version.
}
