% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{amphibians_threatened}
\alias{amphibians_threatened}
\title{Distribution of threatened amphibian species}
\format{A \code{data.frame} with 229425 observations and 5 variables.}
\description{
Data with x,y coordinates for each amphibian species that is Critically Endangered, Endangered or Vulnerable
}
\details{
This dataset contains the distribution of all amphibian species that are 
Critically Endangered, Endangered or Vulnerable according to the IUCN threat status.
}
