% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSR.R
\name{calcSR}
\alias{calcSR}
\title{Calculate species richness of species presence raster stack}
\usage{
calcSR(species_names = NA, path = getwd(), filename = NA, ...)
}
\arguments{
\item{species_names}{List with species names}

\item{path}{Path to location of raster files}

\item{filename}{Specify filename of output}

\item{...}{Additional arguments for saving raster file}
}
\value{
raster layer with species richness of files provided
}
\description{
Read raster files of multiple species to calculate the species richness
With special implementation for calculating the species richness of many species
}
\examples{
\dontrun{
calcSR()
}
}
