% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{gard_reptiles}
\alias{gard_reptiles}
\title{Latin names of Reptile Species}
\format{A \code{data.frame} with 10064 observations and 8 variables.}
\description{
Latin names of Reptile Species
}
\details{
This dataset contains all Latin Names of Reptiles
that were derived from GARD ranges.
}
