% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{gard_reptiles_dist}
\alias{gard_reptiles_dist}
\title{Distribution of reptiles at 0.5 degree spatial resolution}
\format{
A \code{data.frame} with 1001518 observations and 3 variables.
}
\description{
Data with x,y coordinates for each reptile species (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all reptile species (on a 0.5 degree grid)
according to the GARD ranges, version 1.1.
}
