% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{gard_reptiles_dist_smallrange}
\alias{gard_reptiles_dist_smallrange}
\title{Distribution of small ranging reptiles}
\format{
A \code{data.frame} with 15960 observations and 5 variables.
}
\description{
Data with x,y coordinates for each reptile species that has less than 10 occurrences (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all reptile species that 
have a small range (less than 10 occurrences on a 0.5 degree grid)
according to the GARD range maps, version 1.1.
}
