% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBirdLife.R
\name{getBirdLife}
\alias{getBirdLife}
\title{Download range maps of birds from BirdLife International}
\usage{
getBirdLife(user = NA, password = NA, path = getwd())
}
\arguments{
\item{user}{E-mail address you used to login at the IUCN Website}

\item{password}{Password you used to login at the IUCN Website}

\item{path}{Path to download location}
}
\value{
location of data files
}
\description{
This function automatically downloads the BirdLife International
range maps for the world's bird species, 
given that an existing user name and password are provided.
}
\examples{
getBirdLife()
}
