% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIUCN.R
\name{getIUCN}
\alias{getIUCN}
\title{Download IUCN range maps for a given taxa or group}
\usage{
getIUCN(taxa, group, user, password, path = getwd())
}
\arguments{
\item{taxa}{character, Name of taxa to download}

\item{group}{chracter, Name of subgroup, in case you are interested in a subset of species}

\item{user}{E-mail address you used to login at the IUCN Website}

\item{password}{Password you used to login at the IUCN Website}

\item{path}{Path to download location}
}
\value{
location of data files
}
\description{
This function automatically downloads IUCN range maps of a given taxa or group
to a specified location.
}
\examples{
\dontrun{
getIUCN(taxa="Reptiles", group="Chameleons")

getIUCN(group="Chameleons")
}
}
