% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRangeArea.R
\name{getRangeArea}
\alias{getRangeArea}
\title{Obtain range area from species shapefiles}
\usage{
getRangeArea(
  dsn = paste0(getwd(), "IUCN/AMPHIBIANS.shp"),
  id = "binomial",
  seasonal = NA,
  origin = NA,
  presence = NA,
  name_split = c(1, 2),
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{dsn}{Path to one shapefile with multiple polygons or a list of files}

\item{id}{\code{character} Character specifying the id column}

\item{seasonal}{\code{integer} ...}

\item{origin}{\code{integer}}

\item{presence}{\code{integer}}

\item{name_split}{\code{integer} Specifies which splits to use, default is c(1,2).}

\item{crs}{\code{character}}
}
\value{
\code{data.frame} with the range area of each species
}
\description{
Rasterize shapefile of multiple polygons to 
individual global rasters with a specific resolution
}
\examples{
\dontrun{
rasterizeIUCN()
}
}
