% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{odonata}
\alias{odonata}
\title{Latin names of Odonata Species}
\format{
A \code{data.frame} with 3127 observations and 10 variables.
}
\description{
Latin names of Odonata Species
}
\details{
This dataset contains all Latin Names of Odonata
that were derived from the IUCN database, version 6.2 (2019).
}
