% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{odonata_dist}
\alias{odonata_dist}
\title{Distribution of odonata at 0.5 degree spatial resolution}
\format{
A \code{data.frame} with 1071837 observations and 3 variables.
}
\description{
Data with x,y coordinates for each odonata species (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all odonata species (on a 0.5 degree grid)
according to the IUCN range maps, version 6.2 (2019).
}
