% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{odonata_dist_eur}
\alias{odonata_dist_eur}
\title{Distribution of Odonata species at 0.11 degree spatial resolution for Europe}
\format{A \code{data.frame} with 2334175 observations and 7 variables.}
\description{
Data with x,y coordinates for each Odonata species (present in 0.11 degree grid cells) in Europe
}
\details{
This dataset contains the distribution of all Odonata species (on a 0.11 degree grid) in Europe
according to the IUCN ranges.
}
