% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{odonata_endemic}
\alias{odonata_endemic}
\title{Names of endemic odonata species}
\format{
A \code{data.frame} with 529 observations and 2 variables.
}
\description{
Data with endemic IUCN odonata species names and their range area
}
\details{
This dataset contains the name and range area of all endemic odonata species. 
Species are classified as endemic if their range area is smaller or equal 
to the range area of the smallest ranging 15% of all species.
Distribution data was derived from IUCN range maps, version 6.2 (2019).
}
