% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeIUCN.R
\name{rasterizeIUCN}
\alias{rasterizeIUCN}
\title{Rasterize Species Data from Shapefile}
\usage{
rasterizeIUCN(
  dsn = paste0(getwd(), "IUCN/AMPHIBIANS.shp"),
  id = "binomial",
  method = "all",
  resolution = 0.5,
  save = TRUE,
  extent = c(-180, 180, -90, 90),
  split = NA,
  name_split = c(1, 2),
  seasonal = NA,
  origin = NA,
  presence = NA,
  getCover = F,
  df = F,
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0",
  path = getwd()
)
}
\arguments{
\item{dsn}{Path to one shapefile with multiple polygons or a list of files}

\item{id}{\code{character} Character specifying the id column}

\item{method}{\code{character} Default is all, which means both lines and areas of the polygons will be rasterized, 
unless sf and fasterize are installed then only the polygons are rasterized.}

\item{resolution}{\code{integer} Resolution in degrees}

\item{save}{\code{logical} Should individual output be stored. Path to output folder can be specified under path}

\item{extent}{\code{extent} Extent of area that we are interested in.}

\item{split}{\code{character} default is NA}

\item{name_split}{\code{integer} Specifies which splits to use, default is c(1,2).}

\item{seasonal}{\code{integer} 1 = Resident, 2 = Breeding Season, 3 = Non-breeding Season, 4 = Passage, 5 =    Seasonal occurence uncertain.}

\item{origin}{\code{integer} 1 = Native, 2 =    Reintroduced, 3 =    Introduced, 4 =    Vagrant, 5 = Origin Uncertain.}

\item{presence}{\code{integer} 1 =    Extant, 2 = Probably Extant, 3 = Possibly Extant, 4 = Possibly Extinct, 5 = Extinct (post 1500), 6 = Presence Uncertain.}

\item{getCover}{\code{logical} Calculate the percentage covered by a polygon rather than the presence of a species}

\item{df}{\code{logical} Store the output as data.frame or not. If df=FALSE output will be stored as .tif files.}

\item{crs}{\code{character} Define the output projection of your data.}

\item{path}{\code{character} Path where individual output files are going to be stored.}
}
\value{
list of raster layers for each \code{id} with the given area \code{shapefile}
}
\description{
Rasterize shapefile of multiple polygons to 
individual global rasters with a specific resolution
}
\examples{
\dontrun{
rasterizeIUCN()
}
}
