% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterizeSp.R
\name{rasterizeSp}
\alias{rasterizeSp}
\title{Rasterize Species Data}
\usage{
rasterizeSp(dsn)
}
\arguments{
\item{dsn}{Path to one shapefile with multiple polygons or a list of files}
}
\value{
list of raster layers for each \code{id} with the given area \code{shapefile}
}
\description{
Rasterize SpatialPointsDataFrame or SpatialPolygonsDataFrame to 
individual global raster with a specific resolution
}
\examples{
\dontrun{
rasterizeSp()
}
}
