% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{package}
\name{rasterSp package}
\alias{rasterSp package}
\alias{rasterSppackage}
\title{R Package to rasterize and summarise IUCN range maps}
\description{
R Package to rasterize and summarise IUCN range maps
}
\references{
https://www.iucnredlist.org/resources/spatial-data-download; 
http://datazone.birdlife.org/species/spcdistPOS; 
https://datadryad.org/resource/doi:10.5061/dryad.83s7k
}
\author{
RS-eco
}
\keyword{package}
