% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{reptiles}
\alias{reptiles}
\title{Latin names of Reptile Species}
\format{A \code{data.frame} with 7066 observations and 10 variables.}
\description{
Latin names of Reptile Species
}
\details{
This dataset contains all Latin Names of Reptiles
that were derived from the IUCN database.
}
