% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{reptiles_endemic}
\alias{reptiles_endemic}
\title{Names of endemic reptile species}
\format{A \code{data.frame} with 1344 observations and 2 variables.}
\description{
Data with endemic IUCN reptile species names and their range area
}
\details{
This dataset contains the name and range area of all endemic reptile species. 
Species are classified as endemic if their range area is smaller or equal to the range area of the smallest ranging 15% of all species.
}
