% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{reptiles_v6.3}
\alias{reptiles_v6.3}
\title{Latin names of Reptile Species}
\format{
A \code{data.frame} with 10361 observations and 10 variables.
}
\description{
Latin names of Reptile Species
}
\details{
This dataset contains all Latin Names of Reptiles
that were derived from the IUCN database, version 6.3 (2022).
}
