% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{reptiles_v6.3_dist}
\alias{reptiles_v6.3_dist}
\title{Distribution of reptiles at 0.5 degree spatial resolution}
\format{
A \code{data.frame} with 2500866 observations and 3 variables.
}
\description{
Data with x,y coordinates for each reptile species (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all reptile species (on a 0.5 degree grid)
according to the IUCN ranges, version 6.3 (2022).
}
