% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{seabird_species}
\alias{seabird_species}
\title{Latin names of all sea bird species}
\format{A \code{data.frame} with 358 observations and 1 variable.}
\description{
Latin names of all bird species that occur in the ocean
}
\details{
This dataset contains all Latin Names of Birds that have part or all of their range within the sea.
}
