% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesData.R
\name{speciesData}
\alias{speciesData}
\title{Create dataframe of species presence from multiple raster files}
\usage{
speciesData(species_names = NA, path = getwd(), filename = NA, ...)
}
\arguments{
\item{species_names}{List with species names}

\item{path}{Path to location of raster files}

\item{filename}{Specify filename of output}

\item{...}{Additional arguments:
na: String used for missing values. Defaults to NA. Missing values will never be quoted; strings with the same value as na will always be quoted.
append: If FALSE, will overwrite existing file. If TRUE, will append to existing file. In both cases, if file does not exist a new file is created.
col_names:    Write columns names at the top of the file?}
}
\value{
raster layer with species richness of files provided
}
\description{
Read raster files of multiple species and
extract only the coordinates, where individuals are present
}
\examples{
\dontrun{
speciesData()
}
}
