% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_birds}
\alias{ter_birds}
\title{Latin names of Bird Species}
\format{A \code{data.frame} with 17835 observations and 8 variable.}
\description{
Latin names of Bird Species
}
\details{
This dataset contains all Latin Names of Birds
that were derived from the Bird Life Database.
}
