% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_birds_endemic}
\alias{ter_birds_endemic}
\title{Names of endemic bird Species}
\format{
A \code{data.frame} with 1956 observations and 2 variables.
}
\description{
Data with endemic terrestrial bird species names and their range area
}
\details{
This dataset contains the name and range area of all endemic terrestrial bird species. 
Species are classified as endemic if their range area is smaller or equal to the range area 
of the smallest ranging 15% of all species.
Distribution data was derived from BirdLife range maps, version 2015.
}
