% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_birds_threatened}
\alias{ter_birds_threatened}
\title{Distribution of threatened bird species}
\format{
A \code{data.frame} with 204060 observations and 5 variables.
}
\description{
Data with x,y coordinates for each bird species that is Critically 
Endangered, Endangered or Vulnerable
}
\details{
This dataset contains the distribution of all bird species that are 
Critically Endangered, Endangered or Vulnerable according to the IUCN threat status.
Distribution data was derived from BirdLife range maps, version 2015.
}
