% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_mammals}
\alias{ter_mammals}
\title{Latin names of Terrestrial Mammal Species}
\format{A \code{data.frame} with 6067 observations and 10 variables.}
\description{
Latin names of Terrestrial Mammal Species
}
\details{
This dataset contains all Latin Names of Terrestrial Mammals 
that were derived from the IUCN database.
}
