% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_mammals_dist}
\alias{ter_mammals_dist}
\title{Distribution of terrestrial mammals at 0.5 degree spatial resolution}
\format{
A \code{data.frame} with 3396275 observations and 3 variables.
}
\description{
Data with x,y coordinates for each terrestrial mammal species (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all terrestrial mammal species (on a 0.5 degree grid)
according to the IUCN range maps, version 2017.
}
