% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_mammals_dist_smallrange}
\alias{ter_mammals_dist_smallrange}
\title{Distribution of small ranging mammal species}
\format{A \code{data.frame} with 3898 observations and 5 variables.}
\description{
Data with x,y coordinates for each mammal species that has less than 10 occurrences (present in 0.5 degree grid cells)
}
\details{
This dataset contains the distribution of all mammal species that have a small range (less than 10 occurrences on a 0.5 degree grid)
according to the IUCN range maps.
}
