% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_mammals_endemic}
\alias{ter_mammals_endemic}
\title{Names of endemic mammal Species}
\format{
A \code{data.frame} with 4852 observations and 2 variables.
}
\description{
Data with endemic mammal species names and their range area
}
\details{
This dataset contains the name and range area of all endemic terrestrial mammal species. 
Species are classified as endemic if their range area is smaller or equal to the 
range area of the smallest ranging 15% of all species.
Distribution data was derived from IUCN range maps, version 2017.
}
