% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{ter_mammals_threatened}
\alias{ter_mammals_threatened}
\title{Distribution of threatened mammal species}
\format{A \code{data.frame} with 169656 observations and 5 variables.}
\description{
Data with x,y coordinates for each mammal species that is Critically Endangered, Endangered or Vulnerable
}
\details{
This dataset contains the distribution of all mammal species that are 
Critically Endangered, Endangered or Vulnerable according to the IUCN threat status.
}
