% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{threat_status_amphibians}
\alias{threat_status_amphibians}
\title{IUCN threat status for each amphibian species of the IUCN range map data}
\format{
A \code{data.frame} with 6997 observations and 10 variables.
}
\description{
Data with the species information of each amphibian and its current IUCN threat status
}
\details{
This dataset contains the name of every amphibian species present in the IUCN range map data 
and their current IUCN threat status.
}
