% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterSp-package.R
\docType{data}
\name{threat_status_mammals}
\alias{threat_status_mammals}
\title{IUCN threat status for each mammal species of the IUCN range map data}
\format{A \code{data.frame} with 5472 observations and 10 variables.}
\description{
Data with the species information of each mammal and its current IUCN threat status
}
\details{
This dataset contains the name of every mammal species present in the IUCN range map data 
and their current IUCN threat status
}
