% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRangeArea.R
\name{getRangeArea}
\alias{getRangeArea}
\title{Obtain range area from species shapefiles}
\usage{
getRangeArea(
  dsn = paste0(getwd(), "IUCN/AMPHIBIANS.shp"),
  id = "binomial",
  name_split = c(1, 2),
  no_species = NA,
  seasonal = NA,
  origin = NA,
  presence = NA,
  make_valid = F,
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0"
)
}
\arguments{
\item{dsn}{Path to one shapefile with multiple polygons or a list of files}

\item{id}{\code{character} Character specifying the id column}

\item{name_split}{\code{integer} Specifies which splits to use, default is c(1,2).}

\item{no_species}{\code{} Range}

\item{seasonal}{\code{integer} ...}

\item{origin}{\code{integer}}

\item{presence}{\code{integer}}

\item{make_valid}{\code{logical} Default to FALSE. If TRUE polygons are tried to be made valid using sf::st_make_valid()}

\item{crs}{\code{character}}
}
\value{
\code{data.frame} with the range area of each species
}
\description{
Rasterize shapefile of multiple polygons to 
individual global rasters with a specific resolution
}
\examples{
\dontrun{
getRangeArea()
}
}
